Move Span to its own crate
We make `Spanned` a trait, so that it's easier to deal with, and we move
it to its own crate where we also define smart pointers on top of
another trait `Length`. The smart pointers add a `offset` to anything,
and in combination with `Length` can make `Spanned` out of anything that
has a declared length.

We also define special accessors for kdl types so that it's easier to
keep track of the current offset.