Refactor bevy-reflect-deser
This commit migrates bevy-reflect-deser to use the new template-kdl
Navigable trait rather than NodeThunk. To be precise, it uses ThunkField
instead of NodeThunk, but it only uses methods on Navigable. Meaning
that making bevy-reflect-deser generic over the serialization type
should be straightforward.

I also moved about a lot of code. This enabled me to mark a majority of
things as private (while, before, it was mostly pub(crate)) This makes
the code much easier to follow.