Add spawning of kdl scenes into the world
This is likely **utterly broken**! But theoretically, this is supposed
to create a scene that will then be added to the world as child of the
entity with the `KdlScene` component. Workflow is as follow:

- `depends::load_scene` reacts to user-added `KdlScene` component,
  - Find the actual file in file system to load, based on the asset_io
    path (this will not work on environments without FileIo loaders)
  - **Synchronously**[1] load the file, if needed, loads declared
    template files the file imports. All files are cached in the
    `KdlInstances` resource, so that the same template export file isn't
    loaded twice.
  - Insert a `KdlInstance` component to the entity in question, the
    content of the component is a key in `KdlInstances` referring to the
    loaded file declared in `KdlScene`
- `load::load_instance` is exclusive system, reads added `KdlInstance`
  and assumes all of what is described in the previous point was done.
  It's very hacky
  - Converts the `DeserEntity` that was stored as `Box<dyn Reflect>`
  - Creates a dummy `World`, Uses `DeserEntity::spawn_hierarchy` to
    populate the world with the scene in question
  - Stores the dummy world, the entity with the `KdlScene` in question
    and the file name in a `Vec`
  - In a second loop, converts the dummy world into a `Scene` and uses
    `Scene::write_to_world_with` to spawn the scene
  - Makes the entity in question parent of all roots of the
    dummy world and adds the `KdlOrigin` component to all entities in
    the scene

Regarding [1]:

The original plan was to use bevy `Asset`, but it is impractical,
since you can only get the handles to loaded assets within a loading
implementation. The initial impl was asynchronous but not dependent
on bevy `Asset`, but intertwining caching and async turned out to
be way too complicated (you need a way to insert into the cache and
read the cache at the same time). So we opted (temporarily) for a
naive synchronous solution.

I cry seeing the code I wrote, but alas, I couldn't continue without
having something I can run to test.